T3ENCBE-104 refactor js * add styling for content tree
