+++ Resources/Private/Frontend/JavaScript/Utils/stringToHTML.js
@@ -1,0 +1,5 @@
+export default function stringToHTML(htmlString) {
+  const dom = document.createElement('div');
+  dom.innerHTML = htmlString;
+  return dom;
+}
